requirements.txt
langchain==0.2.17
langchain-openai==0.1.25
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
numpy>=1.26,<2.0
